purged modules related to fileio and fetching.
They will form a new repo for fileio only in the future.
Updated `setup.py` with the fewer builds and also modifed `EMAIL` to
update to latest.

Removed unnecessary module dependencies in `setup.py` and `requirements.txt`
that were purged with the removal of modules.

renamed seedfinder.pyx to _seedfinder.pyx and submerpool.pyx to _submerpool.pyx